Refactor from_avro
