+++ .pre-commit-config.yaml
@@ -1,0 +1,10 @@
+repos:
+  -   repo: https://github.com/psf/black.git
+      rev: 22.6.0
+      hooks:
+      - id: black
+        args: [ --config=pyproject.toml ]
+  -   repo: https://github.com/pre-commit/mirrors-isort
+      rev: v5.8.0
+      hooks:
+        - id: isort
