+++ src/pydantic_avro/from_avro/avro_to_pydantic.py
@@ -1,8 +1,8 @@
 import json
 from typing import Optional
 
-from pydantic_avro.class_registery import ClassRegistry
-from pydantic_avro.type_handlers import get_pydantic_type
+from pydantic_avro.from_avro.class_registery import ClassRegistry
+from pydantic_avro.from_avro.type_mapping import get_pydantic_type
 
 
 def validate_schema(schema: dict) -> None:
