+++ src/pydantic_avro/to_avro/config.py
@@ -1,0 +1,4 @@
+from pydantic import VERSION as PYDANTIC_VERSION
+
+PYDANTIC_V2 = PYDANTIC_VERSION.startswith("2.")
+DEFS_NAME = "$defs" if PYDANTIC_V2 else "definitions"
